Add storeValue function to javascript.js
Also added classes to operators and equality
buttons in index.html. And added additional
variables to help prevent hitting the equality
sign too early and when users hit the equality
sign repeatedly after an equation is complete.
Modified the displayNum function to clear the
screen for the total and each user input.